// frontend/src/refinedApi/client/models/LigandMetadataResponse.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type LigandMetadataResponse = {
    id: string;
    name: string;
    experiment_id: string;
    smiles_content?: (string | null);
    link?: (string | null);
    image?: (string | null);
    drug_likeness?: (number | null);
    designed_ligand_score?: (number | null);
};